CHORE: test package structure refactoring
