+++ src/test/java/com/example/waggle/commons/annotation/withMockUser/WithMockCustomUser.java
@@ -1,0 +1,14 @@
+package com.example.waggle.commons.annotation.withMockUser;
+
+import org.springframework.security.test.context.support.WithSecurityContext;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+@WithSecurityContext(factory = WithMockCustomUserSecurityContextFactory.class)
+public @interface WithMockCustomUser {
+    String username() default  "member1";
+
+    String password() default "12345678";
+}
