+++ src/test/java/com/example/waggle/commons/annotation/withMockUser/WithMockCustomUserSecurityContextFactory.java
@@ -1,0 +1,21 @@
+package com.example.waggle.commons.annotation.withMockUser;
+
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.test.context.support.WithSecurityContextFactory;
+
+public class WithMockCustomUserSecurityContextFactory implements WithSecurityContextFactory<WithMockCustomUser> {
+    @Override
+    public SecurityContext createSecurityContext(WithMockCustomUser annotation) {
+
+        final SecurityContext securityContext = SecurityContextHolder.createEmptyContext();
+
+        final UsernamePasswordAuthenticationToken authenticationToken
+                = new UsernamePasswordAuthenticationToken
+                (annotation.username(), "password");
+
+        securityContext.setAuthentication(authenticationToken);
+        return securityContext;
+    }
+}
